Add GetDuration() and SetOnTime() to track device usage information
   Add GerDutation() and SetOnTime() to track device usage info.
   If the arrived message was `on`, SetOnTime() will set `On` of
   the target device, and if the message `off` is arrived,
   GetDuration() will be called to calculated active duration of
   that device.

+++ conf/user.go
@@ -18,7 +18,7 @@ type User struct {
 	Phone string `json:"phone"`
 
 	// User usage summary
-	Usage map[string]int `json:"usage"`
+	Usage map[string]float64 `json:"usage"`
 }
 
 type Users struct {
